circle done

+++ src/maths/dot_product.c
@@ -1,6 +1,6 @@
 # include "../../include/maths.h"
 
-int	dot_product(t_vec v1, t_vec v2)
+double	dot_product(t_vec v1, t_vec v2)
 {
 	return ((v1.x * v2.x) + (v1.y * v2.y) + (v1.z * v2.z));
 }
